Fixed a lot of things after merge fuck up on my side, SORRY
I tried to merge all the completed user branches to have a good starting point to create the trip routes, but along the merges I made a mistake and I overwrote a couple of files. I tried to re-create them as original as possible and on my side everything seems to be working after like an hour of testing, but if you encounter any errors please let me know and I will fix it right away.